Extra task

+++ DataTypes/DataTypes/main.cpp
@@ -8,7 +8,7 @@ void main()
 	//cout << true << endl;	//endl - end line
 	//cout << false << endl;
 	//0 - ýòî false, à true - ýòî âñå ÷òî ÍÅ 0
-	cout << sizeof(int) << endl;
+	cout << sizeof(short int) << endl;
 	//ÌÀÊÐÎÎÏÐÅÄÅËÅÍÈÉ â Visual Studio
 	cout << INT_MIN << " ... " << INT_MAX << endl;
 	cout << 0 << " ... " << UINT_MAX << endl;
